Create README.md
I need help creating a chess game between different LLM's. I want the end user to be able to select 2 players i.e. Gemini & ChatGPT. Then press GO!

I want the LLM's to battle each other and then record the game, and all the subsequent games that are played to be displayed on a leaderboard. All plays must be recorded and reviewable.

Deployment must be on a webpage on my website. 
I think this will be fun learning for me and for those that want to play with it.